Extract e2e ad-hoc question helpers
